nls: remove four hundred and sixty untypable Greek compose sequences
None of the keyboard layouts in xkeyboard-config contains code points
in the range U+1F00..U+1FFF (the Greek Extended block), so none of the
removed sequences could be typed on an existing layout.

(Of course it is possible that some people have their own private layout
that contains some of the Greek Extended characters, but if they do,
then they can also define private compose sequences, when needed.)

Modern Greek does not use these characters with polytonic diacritics
anyway -- they are useful only for typing traditional Greek.

The useless sequences were removed with this command:

  sed -i '/<U1F..>/d' nls/en_US.UTF-8/Compose.pre

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>