nls: make the Amharic compose sequences use the dead-vowel symbols
Last year, commit 488b156fe2 changed mistaken ASCII characters to the
code points that were actually used in the Ethiopian keyboard layout.

But those code points were mistaken too: the intention was not to use
the code point U+FE68 in the layout but the raw symbol code 0xfe68 --
and so on for the other codes.  The symbols that these raw codes were
intended to refer to were `dead_a` and friends.  But these dead vowels
were added under different codes: 0xfe80 to 0xfe8b.

So, correct the Amharic compose sequences to make use of the dead-vowel
symbols that were specifically added for them fifteen years ago.

(As soon as this is merged, xkeyboard-config wil be changed too to use
the dead vowels in the Ethiopian layout.)

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>