nls: use a slash instead of a combining solidus in compose sequences
First: combining diacritics like the combining long solidus (`U+0338`)
are not meant to be used in compose sequences.  Second: there is just
one layout in xkeyboard-config that contains the `U0338` character:
the deprecated/obsolete German T3 layout.  So, practically speaking,
these compose sequences with `U0338` were untypable.  So, use a slash
instead, that almost all layouts have.  This does require that the
sequence `<Compose> <less> <slash>` changes its meaning from backslash
to not-less-than (`≮`).  This seems like an acceptable sacrifice, as
the sequence `<Compose> <slash> <slash>` is a faster/easier sequence
for the backslash and most layouts contain a backslash already anyway.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>