nls: remove four unobvious and redundant compose sequences for ¥ (yen)
Composing ¥ with Y and a minus sign seems to have been added in analogy
to composing £ with L and a minus sign.  But ¥ clearly has a double line
through it, so using the equals sign for this is far more logical (and
those compose sequences of course exist).  Also, L plus an equals sign
produces ₤ (lira), not £ (sterling).  So... make these sequences more
consistent and allow composing ¥ only with Y/y plus an equals sign.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>