nls: harmonize the comments for compose sequences with combining accents
Replace the "WITH" with "plus" and lowercase the "AND" in the comments
for sequences with combining accents, to make it slightly clearer that
the resulting string consists of multiple code points.  Also, use the
word "COMBINING" in the names of the accents, so that these sequences
can be easily grepped, and drop the redundant word "ACCENT".

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>