nls: change result of sequence `<C> <bar>` from `¢` (CENT) to `₵` (CEDI)
As a cent is a small coin, it makes no sense to use an uppercase letter
to compose the `¢` symbol -- having four sequences with a lowercase `c`
plus a `bar` or a `slash` available for composing `¢` should be enough.

Use the sequence `<C> <bar` (with the uppercase `C`) for composing `₵`
(the CEDI SIGN) instead.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>